week2/requirements-livekit.txt
python-dotenv>=1.0.0
livekit-agents>=0.12.0
livekit-plugins-openai>=0.10.0
livekit-plugins-deepgram>=0.7.0
livekit-plugins-elevenlabs>=0.8.0
livekit-plugins-silero>=0.7.0
livekit>=0.18.0
aiohttp>=3.9.0
